TodoProject/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-multipart
python-jose
pydantic[email] 